feat: add audio file validation and security checks for episode creation endpoint
